requirements.txt

alembic==1.16.4
blinker==1.9.0
brotli==1.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
cssselect2==0.8.0
dnspython==2.4.2
email-validator==1.2.1
expat
flask==3.1.0
flask-login==0.6.3
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
fonttools==4.59.0
greenlet==3.2.3
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.2
pillow==11.3.0
pip==25.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
setuptools==78.1.1
sqlalchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
tk
typing-extensions==4.14.1
tzdata
vc
weasyprint==66.0
webencodings==0.5.1
werkzeug==3.1.3
wheel==0.45.1
wtforms==3.2.1
zopfli==0.2.3.post1
python-dateutil
python-dotenv